requirements.txt
requests==2.31.0
anthropic==0.25.0
python-dotenv==1.0.0
sendgrid==6.11.0
pyyaml==6.0.1
streamlit==1.32.0
pandas==2.2.0
plotly==5.19.0
rich==13.7.0